Add Pearce's strongly connected components algorithm
See graph/scc.txt for details on the algorithm.

The space complexity of Tarjan's then 33-years old SCC algorithm was
improved in 2005 by David Pearce. [1] The worst case space complexity of
the algorithm was reduced from O(v(2 + 5w)) to O(3vw) by eliminating the
visit-index table and interleaving the DFS and backtracking stacks into
a single O(vw) data structure.

Tim Leslie added an iterative implementation of Pearce's algorithm
to SciPy in 2013. [2] Instead of using continuations to convert the
recursive calls, Leslie eliminated the need for any continuations by
creating a stack that promotes existing items in the stack to the top in
constant time. This data structure takes O(2vw) space, but eliminates
the need for a continuation variable per vertex (O(vw)), thus preserving
Pearce's original O(3vw) worst case space complexity.

Leslie's implementation does not utilize the call stack and is also very
easy to read and implement, so it is the algorithm that is implemented
here.

Many thanks to @DavePearce and @timleslie for their work!

[1]: http://homepages.ecs.vuw.ac.nz/~djp/files/IPL15-preprint.pdf
[2]: http://www.timl.id.au/SCC